tests/test-read: provide progress updates and improve continuous test bad sample info output
Every second a character should be printed on the screen so users can
tell more easily of the test got hung for any reason. Also, now the bad
samples are collected, checked, and displayed at the end of each session
instead of during. This makes the output info have a higher SNR ratio
rather then printing a lot of excess gtest formatting.