Device: support using fallback values for unbalanced read() in noncontinuous mode
In noncontinuous mode, if more samples are requested via read() than
have been written to a channel, the most recent value written to the
channel will be used to fill out the request.

Note that this isn't enabled in continuous mode and only is triggered if
something has already been written to the channel during the current
run. If samples are requested from a channel in SVMI or SIMV modes with
no values having been written yet during the current session run, a
buffer write timeout will still be thrown.