pysmu.libsmu: Device: add flash_firmware() for individual devices
Instead of using the session object and determining which device to
flash under a multi-device session, now each device object has it's own
flash_firmware method.

Note that this only verifies that at least the related device is
flashed. If any other devices happen to be connected to the system and
sitting in SAM-BA mode, currently they'll be flashed as well.