pysmu.libsmu: split Device into session and non-session classes
This allows for objects to only include methods which will work for
their specific case. For example, devices not added to a session won't
be able to call any methods that deal with streaming data.

Previously devices from session.available_devices (which are accessible
but not in a session) could be used to try to run methods which would
fail on them.